Move auth related functions to separate submodule
