+++ fastagency/helpers.py
@@ -1,20 +1,18 @@
-import hashlib
 import json
-import secrets
-import string
 import uuid
 from typing import Any, Dict, List, Optional, Tuple, Type, TypeVar, Union
 from uuid import UUID
 
 from asyncer import asyncify
 from fastapi import BackgroundTasks, HTTPException
-from pydantic import BaseModel
 
 from fastagency.saas_app_generator import (
     InvalidFlyTokenError,
     InvalidGHTokenError,
     SaasAppGenerator,
 )
+
+from .auth_token.auth import create_deployment_auth_token
 
 # from fastagency.app import add_model
 from .db.helpers import find_model_using_raw, get_db_connection, get_user
@@ -230,72 +228,3 @@ async def create_autogen(
     model = await get_model_by_ref(model_ref)
 
     return await model.create_autogen(model_id=model_id, user_id=user_id, **kwargs)
-
-
-def generate_auth_token(length: int = 32) -> str:
-    characters = string.ascii_letters + string.digits
-    token = "".join(secrets.choice(characters) for _ in range(length))
-    return token
-
-
-def hash_auth_token(token: str) -> str:
-    # Generate a random salt
-    salt = secrets.token_bytes(16)
-
-    # Combine salt and token
-    salted_token = salt + token.encode("utf-8")
-
-    # Use SHA-256 for hashing
-    hashed_token = hashlib.sha256(salted_token).hexdigest()
-
-    # Return the salt and hashed token
-    return salt.hex() + ":" + hashed_token
-
-
-def verify_auth_token(token: str, stored_hash: str) -> bool:
-    if ":" not in stored_hash:
-        return False
-    # Split the stored hash into salt and hash
-    salt, hash_value = stored_hash.split(":")
-
-    # Convert salt back to bytes
-    salt_bytes = bytes.fromhex(salt)
-
-    # Combine salt and token
-    salted_token = salt_bytes + token.encode("utf-8")
-
-    # Hash the salted token
-    computed_hash = hashlib.sha256(salted_token).hexdigest()
-
-    # Compare the computed hash with the stored hash
-    return computed_hash == hash_value
-
-
-class DeploymentAuthToken(BaseModel):
-    auth_token: str
-
-
-async def create_deployment_auth_token(
-    user_uuid: str, deployment_uuid: str
-) -> DeploymentAuthToken:
-    user = await get_user(user_uuid=user_uuid)
-    deployment = await find_model_using_raw(model_uuid=deployment_uuid)
-
-    if user["uuid"] != deployment["user_uuid"]:
-        raise HTTPException(
-            status_code=403, detail="User does not have access to this deployment"
-        )
-
-    auth_token = generate_auth_token()
-    hashed_token = hash_auth_token(auth_token)
-
-    async with get_db_connection() as db:
-        await db.authtoken.create(  # type: ignore[attr-defined]
-            data={
-                "uuid": str(uuid.uuid4()),
-                "deployment_uuid": deployment_uuid,
-                "auth_token": hashed_token,
-            }
-        )
-
-    return DeploymentAuthToken(auth_token=auth_token)
